Create a new tmux session.

Create a new tmux session.

new_session(name = NULL, prompt = NULL, window_name = NULL,
  start_directory = NULL, width = 80, height = 24, detached = TRUE,
  shell_command = NULL)

Arguments

name
String to be used as session name. If NULL (default), the name of the session is determined by tmux, which is the next unused integer (starting at 0).

prompt
String containing a regular expression that matches all relevant prompts.

window_name
String to be used as window name.

start_directory
String indicating the working directory this session is run in.

width
Numeric indicating width of inital window. Default 80 characters.

height
Numeric indicating height of initial window. Default 24 lines. By default, tmux uses one status line, so the effective height is decreased by one line.

detached
Logical. Default TRUE. If FALSE, the R interpreter waits for the session to be killed.

shell_command
String specifying the system command to be invoked when creating the session.

Value

A tmuxr_session.

Examples

new_session()

#> tmuxr session 0: 1 windows (created Sun Aug 27 12:54:53 2017) [80x23]

# Start a session running Bash
s <- new_session(shell_command = "PS1='$ ' bash",
                 prompt = prompts$bash)